Working delete menu
